refactor(rust): move `TcpRouter` into its own file
